+++ irs-api/src/test/java/org/eclipse/tractusx/irs/aaswrapper/job/delegate/RelationshipDelegateTest.java
@@ -191,6 +191,7 @@ class RelationshipDelegateTest {
     @Test
     void shouldCatchUsagePolicyExceptionAndPutTombstone() throws EdcClientException {
         // given
+        final String businessPartnerNumber = "BPNL000000011111";
         final ItemContainer.ItemContainerBuilder itemContainerWithShell = ItemContainer.builder()
                                                                                        .shell(shell("", shellDescriptor(
                                                                                                List.of(submodelDescriptorWithDspEndpoint(
@@ -198,7 +199,7 @@ class RelationshipDelegateTest {
                                                                                                        "address")))));
 
         // when
-        when(submodelFacade.getSubmodelPayload(any(), any(), any(), any())).thenThrow(new UsagePolicyException("itemId", null));
+        when(submodelFacade.getSubmodelPayload(any(), any(), any(), any())).thenThrow(new UsagePolicyException("itemId", null, businessPartnerNumber));
         when(connectorEndpointsService.fetchConnectorEndpoints(any())).thenReturn(List.of("connector.endpoint.nl"));
         final ItemContainer result = relationshipDelegate.process(itemContainerWithShell, jobParameter(),
                 new AASTransferProcess(), createKey());
@@ -207,6 +208,7 @@ class RelationshipDelegateTest {
         assertThat(result).isNotNull();
         assertThat(result.getTombstones()).hasSize(1);
         assertThat(result.getTombstones().get(0).getCatenaXId()).isEqualTo("itemId");
+        assertThat(result.getTombstones().get(0).getBusinessPartnerNumber()).isEqualTo(businessPartnerNumber);
         assertThat(result.getTombstones().get(0).getProcessingError().getProcessStep()).isEqualTo(
                 ProcessStep.USAGE_POLICY_VALIDATION);
     }