+++ CHANGELOG.md
@@ -7,7 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - DigitalTwinRegistryCreateShellService in irs-registry-client for creating shells in DTR directly
+- POST /management/v2/contractagreements/request and GET /management/v2/contractagreements/{contractAgreementId}/negotiation to irs-edc-client lib
 
+### Changed
+- EDC client handles multiple Digital Twin Registries and Digital Twins now #395
+- Change logo of irs
+- Added 'businessPartnerNumber' field to Tombstone model. This will be filled only when UsagePolicyValidation tombstone is being created.
+
+### Fixed
+- Update to Spring Boot 3.1.9 to fix CVE's.
+- Fixed a case where Policy validation could result in a NullPointerException if either orConstraint or andConstraint of
+  accepted policies were null
 
 ## [4.6.0] - 2024-02-20
 ### Added
@@ -25,10 +35,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed Log4J configuration
 - Fix tavern and cucumber tests
 
+## [4.5.2] - 2024-02-22
+## Changed
+- Updated IRS OpenAPI version to 4.5.2
+
 ## [4.5.1] - 2024-02-20
 ### Changed
 - Moved Docker notice to separate file #425
-
 
 ## [4.5.0] - 2024-02-07
 ### Added
@@ -537,7 +550,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Select Aspects you need**  You are able to select the needed aspects for which you want to collect the correct endpoint information.
 
 [Unreleased]: https://github.com/eclipse-tractusx/item-relationship-service/compare/4.6.0...HEAD
-[4.6.0]: https://github.com/eclipse-tractusx/item-relationship-service/compare/4.5.1...4.6.0
+[4.6.0]: https://github.com/eclipse-tractusx/item-relationship-service/compare/4.5.2...4.6.0
+[4.5.2]: https://github.com/eclipse-tractusx/item-relationship-service/compare/4.5.1...4.5.2
 [4.5.1]: https://github.com/eclipse-tractusx/item-relationship-service/compare/4.5.0...4.5.1
 [4.5.1]: https://github.com/eclipse-tractusx/item-relationship-service/compare/4.5.0...4.5.1
 [4.5.0]: https://github.com/eclipse-tractusx/item-relationship-service/compare/4.4.0...4.5.0